pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fim-agent"
version = "0.1.0"
description = "LLM-powered Agent Runtime with dynamic DAG planning and concurrent execution"
readme = "README.md"
license = "LicenseRef-FIM-Agent"
requires-python = ">=3.11"
authors = [
    { name = "fim-ai", email = "dev@fim.ai" },
]
keywords = ["agent", "llm", "dag", "rag", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: Other/Proprietary License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "openai>=1.0",
    "httpx>=0.27",
    "pydantic>=2.0",
    "python-multipart>=0.0.22",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=6.0",
    "ruff>=0.8",
    "mypy>=1.13",
]
web = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "python-dotenv>=1.0",
    "sqlalchemy[asyncio]>=2.0",
    "aiosqlite>=0.20",
    "alembic>=1.14",
    "bcrypt>=4.0",
    "pyjwt>=2.0",
]
mcp = [
    "mcp>=1.0",
]

[project.urls]
Homepage = "https://github.com/fim-ai/fim-agent"
Repository = "https://github.com/fim-ai/fim-agent"
Issues = "https://github.com/fim-ai/fim-agent/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/fim_agent"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "RUF"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
